Inverted logic for determining RTL
This solves a problem with using `MGSwipeTableCell` through `carthage`
in extensions caused by API limitations.

Now on iOS 8 instead of determining if RTL should be used, it has to be
set manually beforehand. For apps it uses the old method, for extensions
it can be either guessed based on user preferred languages or set
manually.