Set and restore the flipped geometry flag
On a table with flipped geometry (where row zero is the bottom of the table) each of the cells usually also have flipped geometry to ensure the UI renders correctly. The `imageFromView` wasn't taking this into account so I've cached the `geometryFlipped` value on the cell so that we can render the image with the natural geometry.